Update LSM Toolbox to version 4.00c

+++ src-plugins/LSM_Toolbox/org/imagearchive/lsm/toolbox/info/ChannelNamesAndColors.java
@@ -15,7 +15,7 @@ public class ChannelNamesAndColors {
 
 	public long[] Reserved;
 
-	public long[] Colors;
+	public int[] Colors;
 	
 	public String[] ChannelNames;
